+++ client/components/index.js
@@ -4,6 +4,6 @@ export {Login} from './auth-form'
 export {default as AllSongs} from './all-songs'
 export {default as SongTraits} from './song-traits'
 export {default as SongTraitSummary} from './song-trait-summary'
-export {default as ReactChart} from './test-bar-chart'
-export {default as RadarChart} from './test-radar-chart'
+export {default as BarChart} from './bar-chart'
+export {default as RadarChart} from './radar-chart'
 export {MeasureDefinitions} from './measure-definitions'
